+++ app/src/main/java/com/kabukabu/driver/features/auth/data/entity/response/UploadCarDocsResponse.kt
@@ -68,7 +68,7 @@ data class UploadCarDocsAutoDebitDates(
 
 @JsonClass(generateAdapter = true)
 data class AverageRating(
-    @Json(name = "value") val value: Int,
+    @Json(name = "value") val value: Double,
     @Json(name = "count") val count: Int
 )
